requirements.txt
numpy>=1.16.0
torch>=1.0.0
torchvision>=0.2.1
opencv-python>=4.0.0
matplotlib>=3.0.0
pillow>=5.3.0
tqdm>=4.30.0
visdom>=0.1.8
scikit-learn>=0.20.0
pandas>=0.23.0